+++ aiven/client/client.py
@@ -3,6 +3,7 @@
 # This file is under the Apache License, Version 2.0.
 # See the file `LICENSE` for details.
 from __future__ import annotations
+from aiven.client.account_client import AivenAccountOrganisationClient
 
 from aiven.client.common_client import AivenCommonClient
 from aiven.client.opensearch_client import AivenOpenSearchClient
@@ -19,5 +20,9 @@ if TYPE_CHECKING:
 UNCHANGED = object()  # used as a sentinel value
 
 
-class AivenClient(AivenCommonClient, AivenOpenSearchClient):
+class AivenClient(
+    AivenCommonClient,
+    AivenAccountOrganisationClient,
+    AivenOpenSearchClient,
+):
     pass
