+++ aiven/client/common_cli.py
@@ -5,7 +5,7 @@
 from __future__ import annotations
 
 from . import argx, units
-from aiven.client import base_client, envdefault
+from aiven.client import base_client
 from aiven.client.base_cli import (
     AivenBaseCLI,
     convert_str_to_value,
@@ -17,7 +17,6 @@ from aiven.client.base_cli import (
     parse_tag_str,
     parse_untag_str,
     REDIS_VALKEY_ACL_ARGS,
-    USER_GROUP_COLUMNS,
 )
 from aiven.client.cliarg import arg
 from aiven.client.common import UNDEFINED
@@ -29,7 +28,7 @@ from aiven.client.speller import suggest
 from collections import Counter
 from decimal import Decimal
 from http import HTTPStatus
-from typing import Any, Callable, Final, IO, Mapping, Optional, Sequence, TypeVar
+from typing import Any, Final, Mapping, Optional, Sequence, TypeVar
 from urllib.parse import urlparse
 
 import errno
@@ -370,153 +369,6 @@ class AivenCommonCLI(AivenBaseCLI):
             disk_desc=disk_desc,
             qual=plan_qual,
         )
-
-    @arg.json
-    @arg("-n", "--name", required=True, help="Name of the account to create")
-    def account__create(self) -> None:
-        """Create new account"""
-        account = self.client.create_account(self.args.name)
-        self.print_response(account, json=self.args.json, single_item=True)
-
-    @arg.json
-    @arg.account_id
-    @arg("-n", "--name", required=True, help="New name for the account")
-    def account__update(self) -> None:
-        """Update an account"""
-        account = self.client.update_account(self.args.account_id, self.args.name)
-        self.print_response(account, json=self.args.json, single_item=True)
-
-    @arg.account_id
-    def account__delete(self) -> None:
-        """Delete an account"""
-        self.client.delete_account(self.args.account_id)
-        print("Deleted")
-
-    @arg.json
-    def account__list(self) -> None:
-        """Lists all current accounts"""
-        accounts = self.client.get_accounts()
-        self.print_response(accounts, json=self.args.json)
-
-    @staticmethod
-    def _parse_auth_config_options(config_cmdline: Sequence[str], config_file: Sequence[str]) -> Mapping[str, str]:
-        options = {}
-        for name_and_value in config_cmdline:
-            if "=" not in name_and_value:
-                raise argx.UserError("Invalid custom value, missing '=': {}".format(name_and_value))
-            name, value = name_and_value.split("=", 1)
-            options[name] = value
-        for name_and_value in config_file:
-            if "=" not in name_and_value:
-                raise argx.UserError("Invalid custom value, missing '=': {}".format(name_and_value))
-            name, filename = name_and_value.split("=", 1)
-            if not os.path.isfile(filename):
-                raise argx.UserError("No such file {!r}".format(filename))
-            with open(filename, encoding="utf-8") as fob:
-                value = fob.read()
-            options[name] = value
-        return options
-
-    @arg.json
-    @arg.account_id
-    def account__team__list(self) -> None:
-        """List account teams"""
-        self.print_response(self.client.list_teams(self.args.account_id), json=self.args.json)
-
-    @arg.json
-    @arg.account_id
-    @arg.team_name
-    def account__team__create(self) -> None:
-        """Create a team within an account"""
-        self.client.create_team(self.args.account_id, self.args.team_name)
-
-    @arg.json
-    @arg.account_id
-    @arg.team_id
-    def account__team__delete(self) -> None:
-        """Delete a team"""
-        self.client.delete_team(self.args.account_id, self.args.team_id)
-
-    @arg.json
-    @arg.account_id
-    @arg.team_id
-    def account__team__user_list(self) -> None:
-        """List team members"""
-        self.print_response(
-            self.client.list_team_members(self.args.account_id, self.args.team_id),
-            json=self.args.json,
-        )
-
-    @arg.json
-    @arg.account_id
-    @arg.team_id
-    @arg.email
-    def account__team__user_invite(self) -> None:
-        """Invite user to join a team"""
-        self.client.add_team_member(self.args.account_id, self.args.team_id, self.args.email)
-
-    @arg.json
-    @arg.account_id
-    @arg.team_id
-    def account__team__user_list_pending(self) -> None:
-        """List pending invitations to a team"""
-        self.print_response(
-            self.client.list_team_invites(self.args.account_id, self.args.team_id),
-            json=self.args.json,
-        )
-
-    @arg.json
-    @arg.account_id
-    @arg.team_id
-    @arg.email
-    def account__team__invite_delete(self) -> None:
-        """Delete pending invite from a team"""
-        self.client.delete_team_invite(self.args.account_id, self.args.team_id, self.args.email)
-
-    @arg.json
-    @arg.account_id
-    @arg.team_id
-    @arg.user_id
-    def account__team__user_delete(self) -> None:
-        """Delete user from a team"""
-        self.client.delete_team_member(self.args.account_id, self.args.team_id, self.args.user_id)
-
-    @arg.json
-    @arg.account_id
-    @arg.team_id
-    def account__team__project_list(self) -> None:
-        """List projects associated to a team"""
-        self.print_response(
-            self.client.list_team_projects(self.args.account_id, self.args.team_id),
-            json=self.args.json,
-        )
-
-    @arg.json
-    @arg.account_id
-    @arg.team_id
-    @arg.project
-    @arg(
-        "--team-type",
-        required=True,
-        choices=["admin", "developer", "operator", "read_only"],
-        help="Team type (permission level)",
-    )
-    def account__team__project_attach(self) -> None:
-        """Attach team to a project"""
-        self.client.attach_team_to_project(
-            self.args.account_id,
-            self.args.team_id,
-            self.args.project,
-            self.args.team_type,
-        )
-
-    @arg.json
-    @arg.account_id
-    @arg.team_id
-    @arg.project
-    def account__team__project_detach(self) -> None:
-        """Detach team from a project"""
-        self.client.detach_team_from_project(self.args.account_id, self.args.team_id, self.args.project)
 
     @optional_auth
     @arg.project
@@ -4460,88 +4312,6 @@ server_encryption_options:
         sbom_url = {"sbom_report_url": f"{self.client.base_url}{self.client.api_prefix}{download_url}"}
         self.print_response(sbom_url)
 
-    @arg.email
-    @arg("--real-name", help="User real name", required=True)
-    def user__create(self) -> None:
-        """Create a user"""
-        password = self.enter_password(
-            "New aiven.io password for {}: ".format(self.args.email),
-            var="AIVEN_NEW_PASSWORD",
-            confirm=True,
-        )
-        result = self.client.create_user(email=self.args.email, password=password, real_name=self.args.real_name)
-
-        self._write_auth_token_file(token=result["token"], email=self.args.email)
-
-    @arg.json
-    def user__info(self) -> None:
-        """Show current user info"""
-        result = self.client.get_user_info()
-        layout = [["user", "real_name", "state", "token_validity_begin", "projects", "auth"]]
-        self.print_response([result], json=self.args.json, table_layout=layout)
-
-    def _write_auth_token_file(self, token: str, email: str) -> None:
-        with self._open_auth_token_file(mode="w") as fp:
-            fp.write(jsonlib.dumps({"auth_token": token, "user_email": email}))
-            aiven_credentials_filename = fp.name
-        os.chmod(aiven_credentials_filename, 0o600)
-        self.log.info("Aiven credentials written to: %s", aiven_credentials_filename)
-
-    def _open_auth_token_file(self, mode: str = "r") -> IO:
-        auth_token_file_path = self._get_auth_token_file_name()
-        try:
-            return open(auth_token_file_path, mode, encoding="utf-8")
-        except OSError as ex:
-            if ex.errno == errno.ENOENT and mode == "w":
-                aiven_dir = os.path.dirname(auth_token_file_path)
-                os.makedirs(aiven_dir)
-                os.chmod(aiven_dir, 0o700)
-                return open(auth_token_file_path, mode, encoding="utf-8")
-            raise
-
-    def _remove_auth_token_file(self) -> None:
-        try:
-            os.unlink(self._get_auth_token_file_name())
-        except OSError:
-            pass
-
-    def _get_auth_token_file_name(self) -> str:
-        default_token_file_path = os.path.join(envdefault.AIVEN_CONFIG_DIR, "aiven-credentials.json")
-        return os.environ.get("AIVEN_CREDENTIALS_FILE") or default_token_file_path
-
-    def _get_auth_token(self) -> str | None:
-        token = self.args.auth_token
-        if token:
-            return token
-
-        try:
-            with self._open_auth_token_file() as fp:
-                return jsonlib.load(fp)["auth_token"]
-        except OSError as ex:
-            if ex.errno == errno.ENOENT:
-                return None
-            raise
-
-    def pre_run(self, func: Callable[[], int | None]) -> None:
-        self.client = self.client_factory(
-            base_url=self.args.url,
-            show_http=self.args.show_http,
-            request_timeout=self.args.request_timeout,
-        )
-        # Always set CA if we have anything set at the command line or in the env
-        if self.args.auth_ca is not None:
-            self.client.set_ca(self.args.auth_ca)
-        if func == self.user__create:
-            # "user create" doesn't use authentication (yet)
-            return
-
-        if not getattr(func, "no_auth", False):
-            auth_token = self._get_auth_token()
-            if auth_token:
-                self.client.set_auth_token(auth_token)
-            elif not getattr(func, "optional_auth", False):
-                raise argx.UserError("not authenticated: please login first with 'avn user login'")
-
     @arg.json
     @arg.project
     def credits__list(self) -> None:
@@ -4928,109 +4698,6 @@ server_encryption_options:
         self.print_response(oauth2_clients, json=self.args.json, table_layout=table_layout)
 
     @arg.json
-    @arg.account_id
-    @arg("--oauth2-client-id", help="OAuth2 client id", required=True)
-    def account__oauth2_client__get(self) -> None:
-        """Get an OAuth2 client configuration."""
-
-        oauth2_client = self.client.get_oauth2_client(self.args.account_id, self.args.oauth2_client_id)
-        table_layout = ["client_id", "name", "description"]
-        self.print_response(oauth2_client, json=self.args.json, single_item=True, table_layout=table_layout)
-
-    @arg.json
-    @arg.account_id
-    @arg("--oauth2-client-id", help="OAuth2 client id", required=True)
-    @arg("-n", "--name", required=True, help="OAuth2 application name")
-    @arg(
-        "-d",
-        "--description",
-        required=False,
-        help="App description",
-    )
-    def account__oauth2_client__update(self) -> None:
-        """Get an OAuth2 client configuration."""
-
-        oauth2_client = self.client.update_oauth2_client(
-            account_id=self.args.account_id,
-            client_id=self.args.oauth2_client_id,
-            name=self.args.name,
-            description=self.args.description,
-        )
-        table_layout = ["client_id", "name", "description"]
-        self.print_response(oauth2_client, json=self.args.json, single_item=True, table_layout=table_layout)
-
-    @arg.json
-    @arg.account_id
-    @arg("--oauth2-client-id", help="OAuth2 client id", required=True)
-    def account__oauth2_client__delete(self) -> None:
-        """Remove an OAuth2 client."""
-
-        self.client.delete_oauth2_client(self.args.account_id, self.args.oauth2_client_id)
-
-    @arg.json
-    @arg.account_id
-    @arg("--oauth2-client-id", help="OAuth2 client id", required=True)
-    def account__oauth2_client__redirect_list(self) -> None:
-        """List OAuth2 client redirects."""
-
-        oauth2_client_redirects = self.client.list_oauth2_client_redirects(self.args.account_id, self.args.oauth2_client_id)
-        table_layout = ["redirect_id", "redirect_uri"]
-        self.print_response(oauth2_client_redirects, json=self.args.json, table_layout=table_layout)
-
-    @arg.json
-    @arg.account_id
-    @arg("--oauth2-client-id", help="OAuth2 client id", required=True)
-    @arg("--redirect-uri", help="Redirect URI")
-    def account__oauth2_client__redirect_create(self) -> None:
-        """Add an allowed redirect URI to an OAuth2 client."""
-
-        redirect = self.client.create_oauth2_client_redirect(
-            self.args.account_id, self.args.oauth2_client_id, self.args.redirect_uri
-        )
-
-        table_layout = ["redirect_id", "redirect_uri"]
-        self.print_response(redirect, json=self.args.json, table_layout=table_layout)
-
-    @arg.json
-    @arg.account_id
-    @arg("--oauth2-client-id", help="OAuth2 client id", required=True)
-    @arg("--redirect-uri-id", help="Redirect URI id", required=True)
-    def account__oauth2_client__redirect_delete(self) -> None:
-        """Add an allowed redirect URI to an OAuth2 client."""
-
-        self.client.delete_oauth2_client_redirect(
-            self.args.account_id, self.args.oauth2_client_id, self.args.redirect_uri_id
-        )
-
-    @arg.json
-    @arg.account_id
-    @arg("--oauth2-client-id", help="OAuth2 client id", required=True)
-    def account__oauth2_client__secret_list(self) -> None:
-        """List OAuth2 client secrets."""
-
-        oauth2_client_secrets = self.client.list_oauth2_client_secrets(self.args.account_id, self.args.oauth2_client_id)
-        table_layout = ["secret_id", "secret_suffix"]
-        self.print_response(oauth2_client_secrets, json=self.args.json, table_layout=table_layout)
-
-    @arg.json
-    @arg.account_id
-    @arg("--oauth2-client-id", help="OAuth2 client id", required=True)
-    def account__oauth2_client__secret_create(self) -> None:
-        """List OAuth2 client secrets."""
-
-        secret = self.client.create_oauth2_client_secret(self.args.account_id, self.args.oauth2_client_id)
-        table_layout = ["secret_id", "secret_suffix", "secret"]
-        self.print_response(secret, json=self.args.json, single_item=True, table_layout=table_layout)
-
-    @arg.json
-    @arg.account_id
-    @arg("--oauth2-client-id", help="OAuth2 client id", required=True)
-    @arg("--secret-id", help="Client secret id")
-    def account__oauth2_client__secret_delete(self) -> None:
-        """List OAuth2 client secrets."""
-
-        self.client.delete_oauth2_client_secret(self.args.account_id, self.args.oauth2_client_id, self.args.secret_id)
-
     @arg.project
     @arg.service_name
     def service__flink__list_applications(self) -> None:
@@ -5314,258 +4981,6 @@ server_encryption_options:
                     file_object=f,
                 ),
             )
-
-    @arg.json
-    @arg("name", help="Name of the organization to create")
-    @arg.force
-    def organization__create(self) -> None:
-        """Create new organization"""
-        if not self.args.force:
-            confirmation_result = self.confirm(
-                "Settings like billing details and authentication methods \
-    cannot be shared across multiple organizations.\
-    \nWhen you create a new organization, you must configure each of these settings manually.\
-    \n\nTo use your current settings, create an organizational unit within this organization instead.\
-    \n\nI understand and want to create a new organization (y/N)? "
-            )
-
-            if not confirmation_result:
-                raise argx.UserError("Aborted")
-
-        organizations = self.client.create_account(self.args.name)
-        layout = [
-            "organization_id",
-            "account_id",
-            "create_time",
-            "update_time",
-        ]
-        self.print_response(organizations, json=self.args.json, table_layout=layout, single_item=True)
-
-    @arg.json
-    def organization__list(self) -> None:
-        """Lists all current organizations"""
-        organizations = self.client.get_organizations()
-        layout = [
-            "organization_name",
-            "organization_id",
-            "account_id",
-            "create_time",
-            "update_time",
-            "tier",
-        ]
-        self.print_response(organizations, json=self.args.json, table_layout=layout)
-
-    @arg.json
-    @arg.organization_id_positional
-    @arg("-n", "--name", required=True, help="New name for the organization")
-    def organization__update(self) -> None:
-        """Update an organization"""
-        layout = [
-            "organization_name",
-            "organization_id",
-            "account_id",
-            "create_time",
-            "update_time",
-        ]
-        organization = self.client.update_organization(self.args.organization_id, self.args.name)
-        self.print_response(organization, json=self.args.json, single_item=True, table_layout=layout)
-
-    @arg.organization_id_positional
-    @arg.force
-    def organization__delete(self) -> None:
-        """Delete an organization"""
-        if not self.args.force:
-            self.print_boxed(
-                [
-                    "Deleting organization cannot be undone and all data in the organization will be lost!",
-                ]
-            )
-
-        if not self.confirm("Confirm delete (y/N)?"):
-            raise argx.UserError("Aborted")
-
-        self.client.delete_organization(self.args.organization_id)
-        print("Deleted")
-
-    @arg.json
-    @arg.organization_id
-    def organization__user__list(self) -> None:
-        """Lists organization users"""
-        users = self.client.list_organization_users(self.args.organization_id)
-        layout = [
-            "user_info.user_email",
-            "user_info.real_name",
-            "user_info.state",
-            "last_activity_time",
-        ]
-        self.print_response(users, json=self.args.json, table_layout=layout)
-
-    @arg.json
-    @arg.organization_id
-    @arg.email
-    def organization__user__invite(self) -> None:
-        """Invite user to join an organization"""
-        self.client.invite_organization_user(self.args.organization_id, self.args.email)
-
-    @arg.json
-    @arg.organization_id
-    def organization__group__list(self) -> None:
-        """List user groups in an organization"""
-        groups = self.client.list_user_groups(self.args.organization_id)
-        layout = [
-            "user_group_name",
-            "user_group_id",
-            "member_count",
-            "description",
-        ]
-        self.print_response(groups, json=self.args.json, table_layout=layout)
-
-    @arg.json
-    @arg.organization_id
-    @arg.group_id_positional
-    def organization__group__show(self) -> None:
-        """Show the user group details"""
-        members = self.client.get_user_group(self.args.organization_id, self.args.group_id)
-        layout = [
-            "user_info.user_email",
-            "user_info.real_name",
-            "last_activity_time",
-            "user_info.state",
-        ]
-        self.print_response(members, json=self.args.json, table_layout=layout)
-
-    @arg.json
-    @arg.organization_id
-    @arg("name", help="Name of the organization user group to create")
-    @arg("--description", help="Description for the organization user group")
-    def organization__group__create(self) -> None:
-        """Create a user group in an organization"""
-        group = self.client.create_user_group(
-            organization_id=self.args.organization_id,
-            group_name=self.args.name,
-            props={
-                "description": self.args.description or "",
-            },
-        )
-        self.print_response(group, json=self.args.json, table_layout=USER_GROUP_COLUMNS, single_item=True)
-
-    @arg.json
-    @arg.organization_id
-    @arg.group_id_positional
-    @arg("--name", help="Name for the organization user group")
-    @arg("--description", help="Description for the organization user group")
-    def organization__group__update(self) -> None:
-        """Update properties of an organization user group"""
-        group = self.client.update_user_group(
-            organization_id=self.args.organization_id,
-            group_id=self.args.group_id,
-            props={
-                "user_group_name": self.args.name,
-                "description": self.args.description,
-            },
-        )
-        self.print_response(group, json=self.args.json, table_layout=USER_GROUP_COLUMNS, single_item=True)
-
-    @arg.force
-    @arg.organization_id
-    @arg.group_id_positional
-    def organization__group__delete(self) -> None:
-        """Delete organization user group"""
-        if not self.args.force:
-            self.print_boxed(
-                [
-                    "Deleting the user group cannot be undone and all data in the group will be lost!",
-                ]
-            )
-
-        if not self.confirm("Confirm delete (y/N)?"):
-            raise argx.UserError("Aborted")
-
-        self.client.delete_user_group(self.args.organization_id, self.args.group_id)
-
-    def _get_stripe_payment_method_id(self, name: str, number: str, exp_month: int, exp_year: int, cvc: str) -> str:
-        """Obtains payment method identifier from Stripe"""
-        stripe_key = self.client.get_stripe_key()
-        raw_client_secret = self.client.create_payment_method_setup_intent()
-
-        request_payload = {
-            "client_secret": raw_client_secret,
-            "payment_method_data[billing_details][name]": name,
-            "payment_method_data[card][number]": number,
-            "payment_method_data[card][cvc]": cvc,
-            "payment_method_data[card][exp_month]": exp_month,
-            "payment_method_data[card][exp_year]": exp_year,
-            "payment_method_data[type]": "card",
-        }
-
-        client_secret_url_part = "_".join(raw_client_secret.split("_")[:2])
-        response = requests.post(
-            f"https://api.stripe.com/v1/setup_intents/{client_secret_url_part}/confirm",
-            data=request_payload,
-            auth=(stripe_key, ""),
-            timeout=30,
-        )
-
-        if not response.ok:
-            response.raise_for_status()
-        return response.json()["payment_method"]
-
-    @arg.json
-    @arg.organization_id
-    def organization__card__list(self) -> None:
-        """List organization cards"""
-        cards = self.client.list_payment_methods(self.args.organization_id)
-        for card in cards:
-            card["expiration"] = f"{card['exp_month']:02d}/{card['exp_year']}"
-            card["number"] = f"**** **** **** {card['last4']}"
-
-        layout = [
-            "card_id",
-            "name",
-            "number",
-            "expiration",
-        ]
-        self.print_response(cards, json=self.args.json, table_layout=layout)
-
-    @arg.json
-    @arg.organization_id
-    @arg("--cvc", help="Credit card security code", required=True)
-    @arg("--exp-month", help="Card expiration month (1-12)", type=int, required=True)
-    @arg("--exp-year", help="Card expiration year", type=int, required=True)
-    @arg("--name", help="Name on card", required=True)
-    @arg("--number", help="Credit card number", type=int, required=True)
-    def organization__card__create(self) -> None:
-        """Add a credit card"""
-
-        payment_method_id = self._get_stripe_payment_method_id(
-            self.args.name,
-            self.args.number,
-            self.args.exp_month,
-            self.args.exp_year,
-            self.args.cvc,
-        )
-        card = self.client.attach_payment_method(self.args.organization_id, payment_method_id)
-
-        card["expiration"] = f"{card['exp_month']:02d}/{card['exp_year']}"
-        card["number"] = f"**** **** **** {card['last4']}"
-        layout = [
-            "card_id",
-            "name",
-            "number",
-            "expiration",
-        ]
-        self.print_response(card, json=self.args.json, table_layout=layout, single_item=True)
-
-    @arg.json
-    @arg.force
-    @arg.organization_id
-    @arg("card_id", help="Credit card identifier")
-    def organization__card__delete(self) -> None:
-        """Delete organization card"""
-        if not self.confirm("Confirm delete (y/N)?"):
-            raise argx.UserError("Aborted")
-        self.client.delete_organization_card(self.args.organization_id, self.args.card_id)
-        print("Deleted")
 
     @arg.project
     @arg.cloud
